Message Server - now supports multiple clients per queue
    Each named queue now supports multiple clients.  Each client will receive its own queue of messages
that it can pop.  When messages are queued they will be added to the queue for each client.  If no clients are
registered then the requests are not saved.  Clients will timeout after 5.5 minutes of inactivity.

Each client can also clear its own queue if it needs to.  Clients likely want to do this when first rendering
the display.

All other functionality remains the same.